Refactoring to enable programatic usage
