+++ index.js
@@ -1,2 +1,3 @@
 module.exports.Contract = require('./lib/contract');
 module.exports.Joi = require('joi');
+module.exports.validateContracts = require('./lib/validator')
