+++ lib/runner.js
@@ -4,6 +4,7 @@ var FileHound = require('filehound');
 var path = require('path');
 var fs = require('fs');
 var formatter = require('./formatter');
+var validateContracts = require('./validator')
 
 function validateContract(contract, cb) {
   contract.validate(function (err) {
@@ -25,23 +26,7 @@ function validateFiles(files) {
     return process.exit(1);
   }
 
-  async.mapSeries(contracts, validateContract, function (err, results) {
-    var failures = _(results).filter('err').compact().value();
-    var totalCompleted = contracts.length;
-    var totalFailed = failures.length;
-    var totalPassed = totalCompleted - totalFailed;
-    var exitCode = (failures.length > 0) ? 1 : 0;
-
-    formatter.print({
-      results: results,
-      failures: failures,
-      totalCompleted: totalCompleted,
-      totalFailed: totalFailed,
-      totalPassed: totalPassed
-    });
-
-    process.exit(exitCode);
-  });
+  validateContracts(contracts);
 }
 
 module.exports = function (files) {
